+++ packages/ts-example-without-defaults/vite.config.js
@@ -27,7 +27,7 @@ export default defineConfig(({ mode }) => {
           },
         },
       },
-      chunkSizeWarningLimit: 441, // @maxgraph/core
+      chunkSizeWarningLimit: 440, // @maxgraph/core
     },
   };
 });
