+++ packages/ts-example/vite.config.js
@@ -27,7 +27,7 @@ export default defineConfig(({ mode }) => {
           },
         },
       },
-      chunkSizeWarningLimit: 445, // @maxgraph/core
+      chunkSizeWarningLimit: 444, // @maxgraph/core
     },
   };
 });
